Разбиваем фай js на модули
