+++ index.html
@@ -233,6 +233,7 @@
   </template>
   <script src="js/index.js"></script>
   <script src="js/functions.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 
 </html>
